Use types.CommitteeIndex

+++ beacon-chain/sync/validate_sync_committee_message.go
@@ -86,7 +86,7 @@ func (s *Service) validateSyncCommitteeMessage(ctx context.Context, pid peer.ID,
 	// Validate that the validator is in the correct committee.
 	subCommitteeSize := params.BeaconConfig().SyncCommitteeSize / params.BeaconConfig().SyncCommitteeSubnetCount
 	for _, idx := range committeeIndices {
-		subnet := idx / subCommitteeSize
+		subnet := uint64(idx) / subCommitteeSize
 		if strings.HasPrefix(*msg.Topic, fmt.Sprintf(format, digest, subnet)) {
 			isValid = true
 			break
@@ -99,7 +99,7 @@ func (s *Service) validateSyncCommitteeMessage(ctx context.Context, pid peer.ID,
 	// There has been no other valid sync committee signature for the declared `slot`, `validator_index` and `subcommittee_index`.
 	// In the event of `validator_index` belongs to multiple subnets, as long as one subnet has not been seen, we should let it in.
 	for _, idx := range committeeIndices {
-		subnet := idx / subCommitteeSize
+		subnet := uint64(idx) / subCommitteeSize
 		if s.hasSeenSyncMessageIndexSlot(m.Slot, m.ValidatorIndex, subnet) {
 			isValid = false
 		} else {
@@ -139,7 +139,7 @@ func (s *Service) validateSyncCommitteeMessage(ctx context.Context, pid peer.ID,
 	}
 
 	for _, idx := range committeeIndices {
-		subnet := idx / subCommitteeSize
+		subnet := uint64(idx) / subCommitteeSize
 		s.setSeenSyncMessageIndexSlot(m.Slot, m.ValidatorIndex, subnet)
 	}
